use fetch API for readable records
